Add files via upload
there is a problem with the keysearch, i cant figure it out, id love it if you could help w it!

the if statement requires a true/false value to be returned, but keysearch returns a tuple with {value, tuple} if it's not false. Idk how to make that into a true but if you could figure that out id really appreciate it. I tested all the rest of the stuff.